Change background color and remaining stylings etc
Changed background color to a lighter creamy color. Also changed the
corresponding styling/colors to prevent accessibility issues and improve
aesthetics. Added a typing header animation to the header section using
TypeAnimation. Also added responsive design to the portfolio section and
imporoved general responsiveness of the site. Had previously added a
circular cursor that follows mouse for a modern cursor effect but forgot
to explain it in a previous commit so will mention here.